pyproject.toml
[tool.poetry]
name = "cyborg"
version = "0.1.0"
description = ""
authors = ["Ryan <rybergy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"ruamel.yaml" = "^0.17.21"
"discord.py" = "^1.7.3"
pydantic = "^1.9.1"
PyNaCl = "^1.5.0"
pytube = "^12.1.0"
plumbum = "^1.7.2"
async-timeout = "^3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
